Moved leaderboards to single command
