+++ src/main/java/net/javadiscord/javabot/systems/help/commands/HelpCommandHandler.java
@@ -1,9 +1,9 @@
 package net.javadiscord.javabot.systems.help.commands;
 
 import net.javadiscord.javabot.command.DelegatingCommandHandler;
-import net.javadiscord.javabot.systems.help.commands.subcommands.ExperienceLeaderboardSubcommand;
+import net.javadiscord.javabot.systems.commands.subcommands.leaderboard.ExperienceLeaderboardSubcommand;
 import net.javadiscord.javabot.systems.help.commands.subcommands.HelpAccountSubcommand;
-import net.javadiscord.javabot.systems.help.commands.subcommands.ThanksLeaderboardSubcommand;
+import net.javadiscord.javabot.systems.commands.subcommands.leaderboard.ThanksLeaderboardSubcommand;
 
 /**
  * Handler class for all Help Commands.
@@ -15,7 +15,5 @@ public class HelpCommandHandler extends DelegatingCommandHandler {
 	 */
 	public HelpCommandHandler() {
 		this.addSubcommand("account", new HelpAccountSubcommand());
-		this.addSubcommand("thanks-leaderboard", new ThanksLeaderboardSubcommand());
-		this.addSubcommand("experience-leaderboard", new ExperienceLeaderboardSubcommand());
 	}
 }
